requirements.txt
asn1crypto==1.5.1
bit==0.8.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==20.0.0
idna==3.7
imageio==2.34.2
lazy_loader==0.4
networkx==3.3
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pycparser==2.22
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
tifffile==2024.7.2
urllib3==2.2.2
streamlit==1.25.0
face-recognition==1.3.0
pickle==1.0